:truck: Move square placeholder function to grid
