+++ packages/web/src/components/MandalaChart/Grid.tsx
@@ -1,4 +1,7 @@
+import useI18n from '@/hooks/useI18n'
+import DisplayingSquare from './DisplayingSquare'
 import Square from './Square'
+import TRANSLATIONS from './Square.i18n'
 
 interface GridProps {
   wholeGridValues: string[][]
@@ -22,22 +25,60 @@ const Grid = ({
   const isGridValid =
     gridIndex === 4 ? true : wholeGridValues[4][gridIndex] !== ''
   const values = wholeGridValues[gridIndex]
+  const { getTranslation } = useI18n()
+  const translation = getTranslation(TRANSLATIONS)
 
+  const getSquarePlaceHolder = (
+    isCenterGrid: boolean,
+    isCenterSquare: boolean,
+    gridIndex: number,
+    squareIndex: number
+  ) => {
+    if (isCenterGrid && isCenterSquare) return translation('mainGoal')
+    if (isCenterGrid && !isCenterSquare)
+      return `${translation('subGoal')} ${
+        squareIndex < 4 ? squareIndex + 1 : squareIndex
+      }`
+    if (!isCenterGrid && isCenterSquare)
+      return `${translation('subGoal')} ${
+        gridIndex < 4 ? gridIndex + 1 : gridIndex
+      }`
+    return ''
+  }
+
+  const isCenterGrid = gridIndex === 4
   return (
     <div className={`grid grid-cols-3 gap-1 w-max`}>
-      {values.map((value, suqareIndex) => {
+      {values.map((value, squareIndex) => {
+        const isCenterSquare = squareIndex === 4
         const handleDoubleClick =
-          getHandleDoubleClick && getHandleDoubleClick(suqareIndex)
+          getHandleDoubleClick && getHandleDoubleClick(squareIndex)
 
-        return (
+        const placeHolder = getSquarePlaceHolder(
+          isCenterGrid,
+          isCenterSquare,
+          gridIndex,
+          squareIndex
+        )
+
+        return isAIModeOn ? (
+          <DisplayingSquare
+            key={squareIndex}
+            value={value}
+            isGridValid={isGridValid}
+            gridIndex={gridIndex}
+            squareIndex={squareIndex}
+          />
+        ) : (
           <Square
-            key={suqareIndex}
+            key={squareIndex}
             value={value}
             handleGridValue={handleGridValue}
             handleDoubleClick={handleDoubleClick}
             isGridValid={isGridValid}
             gridIndex={gridIndex}
-            squareIndex={suqareIndex}
+            squareIndex={squareIndex}
+            placeHolder={placeHolder}
           />
         )
       })}
